Moving visibles classes into DQM
